rolesテーブルのマイグレーションファイルの修正

+++ backend/database/migrations/2021_05_04_070922_create_roles_table.php
@@ -15,7 +15,7 @@ class CreateRolesTable extends Migration
     {
         Schema::create('roles', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->int('role_id');
+            $table->unsignedBigInteger('role_id');
             $table->string('role_name');
             $table->dateTime('created_at', 0)->useCurrent();
             $table->dateTime('updated_at', 0)->useCurrent();
